Start develop upload error log function

+++ branches/v031/source/App/drv_rtc.h
@@ -20,8 +20,8 @@
 /* Exported types ------------------------------------------------------------*/
 struct RTC_STATUS {
 	//unsigned char update_count ;
-	unsigned int update_timer ;//re-update if > RTC_UPDATE_PERIOD
-	unsigned int prescaler ;  //++ if rtc slow, -- if rtc fast
+	u32 update_timer ;//re-update if > RTC_UPDATE_PERIOD
+	u16 prescaler ;  //++ if rtc slow, -- if rtc fast
 };
 
 struct DATE_TIME
